api-gateway/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29
websockets>=12.0
asyncpg>=0.29
aiokafka>=0.11
lz4>=4.0
redis>=5.0
pydantic>=2.0
structlog==25.5.0
python-dotenv>=1.0
